# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "StopVol API",
        "description": "API pour la déclaration et le suivi des vols d'engins (motos, véhicules, etc.)",
        "contact": {
            "name": "StopVol Support",
            "email": "contact@stopvol.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "StopVol API Server"
        }
    ],
    "paths": {
...
        "/api/profile/complete": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Compléter le profil utilisateur",
                "description": "Complète le profil utilisateur avec les informations personnelles et documents requis",
                "operationId": "4a666cec853f5d35f11cbab03930ea51",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "city",
                                    "district",
                                    "id_type"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nom complet",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "city": {
                                        "description": "Ville de résidence",
                                        "type": "string",
                                        "example": "Ouagadougou"
                                    },
                                    "district": {
                                        "description": "Quartier/Secteur",
                                        "type": "string",
                                        "example": "Secteur 15"
                                    },
                                    "id_type": {
                                        "description": "Type de document d'identité",
                                        "type": "string",
                                        "enum": [
                                            "cnib",
                                            "permis",
                                            "passeport"
                                        ],
                                        "example": "cnib"
                                    },
                                    "profile_picture": {
                                        "description": "Photo de profil",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "id_card_front": {
                                        "description": "Recto du document d'identité",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "id_card_back": {
                                        "description": "Verso du document d'identité",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil complété avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profil complété avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "profile_status": {
                                                    "type": "string",
                                                    "example": "pending_validation"
                                                },
                                                "profile_complete": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Données invalides"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Profil déjà complété",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Le profil est déjà complété"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/update": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Mettre à jour le profil utilisateur",
                "description": "Met à jour les informations du profil utilisateur",
                "operationId": "c7d8a1e02c8bb3754e9ee603f0d817ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nom complet",
                                        "type": "string",
                                        "example": "John Doe Updated"
                                    },
                                    "city": {
                                        "description": "Ville de résidence",
                                        "type": "string",
                                        "example": "Bobo-Dioulasso"
                                    },
                                    "district": {
                                        "description": "Quartier/Secteur",
                                        "type": "string",
                                        "example": "Secteur 20"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profil mis à jour avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe Updated"
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "Bobo-Dioulasso"
                                                },
                                                "district": {
                                                    "type": "string",
                                                    "example": "Secteur 20"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Données invalides"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utilisez le token Bearer obtenu après authentification OTP",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}